invoice-processing-pipeline/processor/requirements.txt
google-cloud-documentai==1.2.1
google-cloud-storage==2.0.0
google-cloud-firestore==2.3.4